Extract WorkerRunner from WorkerCommand
Moves everything WorkerCommand does after CommandHelper::begin() — the
TCP connect, NDJSON handshake and per-file analysis loop — into a
reusable WorkerRunner service. WorkerCommand now only boots and
delegates. This makes the post-boot worker logic callable without a
re-boot, which the pcntl_fork() path will rely on.
